Docker: check if directory already exists

+++ MARIE_AND_BERT/prestart.sh
@@ -1,13 +1,18 @@
 mkdir  -p /app/DATA
 mkdir -p /root/.local/share/ChemDataExtractor/
 
-wget -nc http://159.223.42.53:8080/CrossGraph.zip && unzip CrossGraph.zip -d /app/DATA/ && rm CrossGraph.zip
-wget -nc http://159.223.42.53:8080/Evaluation.zip && unzip Evaluation.zip -d /app/DATA/ && rm Evaluation.zip
-wget -nc http://159.223.42.53:8080/EntityLinking.zip && unzip EntityLinking.zip -d /app/DATA/ && rm EntityLinking.zip
-wget -nc http://159.223.42.53:8080/bert_pretrained.zip && unzip bert_pretrained.zip -d /app/ && rm bert_pretrained.zip
-wget -nc http://159.223.42.53:8080/Dictionaries.zip && unzip Dictionaries.zip -d /app/DATA/ && rm Dictionaries.zip
-wget -nc http://159.223.42.53:8080/cde_models.zip && unzip cde_models.zip -d /root/.local/share/ChemDataExtractor/ && rm cde_models.zip
+if [ -d /app/DATA/CrossGraph ];
 
+then
+  echo "File already exists"
+else
+  wget -nc http://159.223.42.53:8080/CrossGraph.zip && unzip CrossGraph.zip -d /app/DATA/ && rm CrossGraph.zip
+  wget -nc http://159.223.42.53:8080/Evaluation.zip && unzip Evaluation.zip -d /app/DATA/ && rm Evaluation.zip
+  wget -nc http://159.223.42.53:8080/EntityLinking.zip && unzip EntityLinking.zip -d /app/DATA/ && rm EntityLinking.zip
+  wget -nc http://159.223.42.53:8080/bert_pretrained.zip && unzip bert_pretrained.zip -d /app/ && rm bert_pretrained.zip
+  wget -nc http://159.223.42.53:8080/Dictionaries.zip && unzip Dictionaries.zip -d /app/DATA/ && rm Dictionaries.zip
+  wget -nc http://159.223.42.53:8080/cde_models.zip && unzip cde_models.zip -d /root/.local/share/ChemDataExtractor/ && rm cde_models.zip
+fi
 
 #wget http://www.theworldavatar.com/MARIE_DEPLOYMENT/CrossGraph.zip && unzip CrossGraph.zip -d /app/DATA/ && rm CrossGraph.zip
 #wget http://www.theworldavatar.com/MARIE_DEPLOYMENT/Evaluation.zip && unzip Evaluation.zip -d /app/DATA/ && rm Evaluation.zip
